Added a null map check for quest nodes using maps
If I'm understanding this correctly, there should never be a null map when `TestRunInt` or `RunInt` for those specific quest nodes.

However, I somehow managed to achieve a situation where the map is null when running `TestRunInt`, causing a `NullReferenceException` - specifically, any time (even an unrelated) quest was generated each of those quest nodes would throw that specific exception (besides DelayedRitualReward, as it doesn't operate on the map in `TestRunInt` - it only uses it in `RunInt`). It was a heavily modded game, and I was not able to reproduce it a second time - only happening on 1 saved game.

So as I've just explained - it's extremely unlikely that the null map check will be useful for big majority of the players... However, this should provide some extra safety in case something goes really bad, as it happened to me.