Fix grand ball/art exhibit bug
I've introduced a bug with a recent changes. This should fix it.

Basically, by turing `.OrderBy(order).Where(condition).First()` into `.OrderBy(order).First(condition)`, I've made it so that `First` call works on an unordered list first (basically a `.Where()` call) and returns the first correct one, as sorted by `OrderBy()`. Before that the `First()` call was guaranteed to only be called once, only on the ordered and filtered list.