Quests no longer proposes giving honor to temporary pawns
Basically, deserter hideout quest uses a quest transporting player pawns in a shuttle, while also providing temporary pawns to help. Vanilla never combines the 2, so the code for selecting a pawn never expected there to be temporary quest pawns.

The fix here is to create a custom ChoiceLetter that inherits from the one responsible for choosing pawns, but filtering the pawn list to exclude pawns who aren't colonists.